requirements/packaged_requirements.txt
--find-links /usr/local/airflow/plugins
--no-index
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.6.3/constraints-3.10.txt"

apache-airflow-providers-snowflake==4.2.0
apache-airflow-providers-mysql==5.1.1
elasticsearch7==7.17.9
apache-airflow==2.6.3
python-dotenv==1.0.0
boto3==1.26.161
pytest==7.4.0
jupyter==1.0.0
